+++ engine/client/s_stream.c
@@ -76,7 +76,7 @@ float S_GetMusicVolume( void )
 S_StartBackgroundTrack
 =================
 */
-void S_StartBackgroundTrack( const char *introTrack, const char *mainTrack, long position )
+void S_StartBackgroundTrack( const char *introTrack, const char *mainTrack, int position )
 {
 	S_StopBackgroundTrack();
